Refactor getName, getSlug
